+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/Webhook/event_subscribers.yml
@@ -1,0 +1,29 @@
+services:
+    Akeneo\Connectivity\Connection\Infrastructure\Webhook\EventSubscribers\EventsApiRequestsLimitEventSubscriber:
+        arguments:
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Webhook\Service\GetDelayUntilNextRequest'
+            - '%webhook_requests_limit%'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Webhook\Service\Sleep'
+            - '@Akeneo\Connectivity\Connection\Application\Webhook\Service\Logger\ReachRequestLimitLogger'
+            - '@Akeneo\Connectivity\Connection\Application\Webhook\Service\EventsApiDebugLogger'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Webhook\EventsApiDebug\Persistence\ElasticsearchEventsApiDebugRepository'
+        tags:
+            - { name: kernel.event_subscriber }
+
+    Akeneo\Connectivity\Connection\Infrastructure\Webhook\EventSubscribers\EventsApiRequestsLimitIncrementSubscriber:
+        arguments:
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Webhook\Persistence\DbalEventsApiRequestCountRepository'
+        tags:
+            - { name: kernel.event_subscriber }
+
+    Akeneo\Connectivity\Connection\Infrastructure\Webhook\EventSubscribers\EventsApiClearCacheSubscriber:
+        arguments:
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Webhook\Service\CacheClearer'
+        tags:
+            - { name: kernel.event_subscriber }
+
+    Akeneo\Connectivity\Connection\Infrastructure\Webhook\EventSubscribers\EventsApiLoggingSubscriber:
+        arguments:
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Webhook\EventsApiDebug\Persistence\ElasticsearchEventsApiDebugRepository'
+        tags:
+            - { name: kernel.event_subscriber }
